chore: moved history data to inside History folder
